+++ src/views/layouts/checkout.hbs
@@ -3,6 +3,7 @@
     <form class='js-form'>
       {{{ordering}}}
       {{{createPayment}}}
+      {{{backdropMarkup}}}
     </form>
   </div>
 </section>
