+++ src/views/partials/checkout/payment.hbs
@@ -81,7 +81,7 @@
       </div>
       
     </div>
-    <button class="button__checkout" type='submit'>ОФОРМИТЬ ЗАКАЗ</button>
+    <button class="button__checkout" type='submit' data-modal-open >ОФОРМИТЬ ЗАКАЗ</button>
   </div>
 
   <div class='description'>
